sam-finetune/requirements.txt
ipython==7.31.1
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.8.0.74
openpyxl==3.1.2
Pillow==10.1.0
segment_anything==1.0
torch==2.0.1
segment_anything
transformers